Fixed some routing issues
