+++ src/main/react/src/pages/InaccessibleResource.tsx
@@ -1,0 +1,12 @@
+// TODO Add automatic redirect set to timer. And a manual 'back to home' button.
+
+export default function InaccessibleResource() {
+
+  return (
+    <div>
+      <h1>This page should be inaccessible</h1>
+      <p>Sorry, an unexpected error has occurred.</p>
+
+    </div>
+  );
+}
