--- src/main/react/src/service/authRouteService.tsx
@@ -1,4 +1,0 @@
-export async function allowUsersAndAbove(){
-    console.log("asdf");
-    return null;
-}
